\name{cmd.examl}
\alias{cmd.examl}
\title{Produce a single ExaML shell command.}
\usage{
  cmd.examl(indir, infile, outdir = indir,
    prog.parser = PR.EXAML.PARSER,
    prog.starttree = PR.EXAML.STARTTREE,
    args.starttree.seed = 12345, args.starttree.bsid = NA,
    prog.examl = PR.EXAML.EXAML,
    args.examl = "-m GAMMA -D", resume = 1, verbose = 1)
}
\value{
  Character string
}
\description{
  Internal code.
}

