% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.cmd.R
\name{cmd.fasttree.add.bootstrap.nodelabels.to.base.tree}
\alias{cmd.fasttree.add.bootstrap.nodelabels.to.base.tree}
\title{Produce a single FastTree shell command on a bootstrap alignment.}
\usage{
cmd.fasttree.add.bootstrap.nodelabels.to.base.tree(bs.dir, bs.n,
  infile.tree.to.add.labels, outfile,
  bs.pattern = "*_ft.[0-9][0-9][0-9].newick", pr = PR.RAXML)
}
\value{
Character string
}
\description{
Produce a single FastTree shell command on a bootstrap alignment.
}
