% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.cmd.R
\name{cmd.fasttree.many.bootstraps}
\alias{cmd.fasttree.many.bootstraps}
\title{Produce a single FastTree shell command on a bootstrap alignment.}
\usage{
cmd.fasttree.many.bootstraps(infile.fasta, bs.dir, bs.n, outfile,
  pr = PR.FASTTREE, pr.args = "-nt -gtr -gamma",
  prog.bscreate = PR.EXAML.BSCREATE, opt.bootstrap.by = "nucleotide",
  pr.supportadder = PR.RAXML, bs.from = NA, bs.to = NA)
}
\value{
Character string
}
\description{
Produce a single FastTree shell command on a bootstrap alignment.
}
