% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.cmd.R
\name{cmd.fasttree.one.bootstrap}
\alias{cmd.fasttree.one.bootstrap}
\title{Produce a single FastTree shell command on a bootstrap alignment.}
\usage{
cmd.fasttree.one.bootstrap(infile.fasta, bs.id,
  outfile = gsub("\\\\.fa|\\\\.fasta|\\\\.FA|\\\\.FASTA", paste0("_",
  sprintf("\%03d", bs.id), ".newick"), infile.fasta), pr = PR.FASTTREE,
  pr.args = "-nt -gtr -gamma", prog.bscreate = PR.EXAML.BSCREATE,
  opt.bootstrap.by = "nucleotide", check.binary = TRUE)
}
\value{
Character string
}
\description{
Produce a single FastTree shell command on a bootstrap alignment.
}
