% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.cmd.R
\name{cmd.fasttree}
\alias{cmd.fasttree}
\title{Produce a single FastTree shell command.}
\usage{
cmd.fasttree(infile.fasta, outfile = paste(infile.fasta, ".newick", sep =
  ""), pr = PR.FASTTREE, pr.args = "-nt -gtr -gamma",
  check.binary = FALSE)
}
\value{
Character string
}
\description{
Produce a single FastTree shell command.
}
