% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.cmd.R
\name{cmd.jmodeltest}
\alias{cmd.jmodeltest}
\title{Produce a single LSD shell command.}
\usage{
cmd.jmodeltest(infile.fasta, outfile = paste0(infile.fasta,
  ".jmodeltest"), pr = PR.JMODELTEST,
  pr.args = "-f -i -g 4 -s 3 -DT -S NNI -t ML", nproc = 1)
}
\value{
Character string
}
\description{
Produce a single LSD shell command.
}
