% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.cmd.R
\name{cmd.lsd.dates}
\alias{cmd.lsd.dates}
\title{Generate the LSD Dates file.}
\usage{
cmd.lsd.dates(infile.dates, infile.tree, outfile.lsd.dates,
  pr = PR.LSDDATES, run.lsd = FALSE, outfile.lsd = NA, root = NA,
  exclude.missing.dates = FALSE, outfile.tree = NA, ali.len = NA)
}
\value{
Character string
}
\description{
Generate the LSD Dates file.
}
