% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.cmd.R
\name{cmd.lsd}
\alias{cmd.lsd}
\title{Produce a single LSD shell command.}
\usage{
cmd.lsd(infile.tree, infile.dates, ali.nrow, outfile = infile.tree,
  PR.LSD = "lsd", pr.args = "-v 2 -c -b 10 -r as")
}
\value{
Character string
}

