% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.cmd.R
\name{cmd.multi2di}
\alias{cmd.multi2di}
\title{Produce a shell command to convert multi-furcating trees to binary trees}
\usage{
cmd.multi2di(infile, outfile, seed = 42, pr = PR.MULTI2DI)
}
\value{
Character string
}
\description{
Produce a shell command to convert multi-furcating trees to binary trees
}
