% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.cmd.R
\name{cmd.mvr}
\alias{cmd.mvr}
\title{Produce a single shell command to run MVR.}
\usage{
cmd.mvr(infile, outfile, prog = PR.MVR, method = "MVR",
  complete.distance.matrix = FALSE, name.gd.col = "GD")
}
\value{
Character string
}
\description{
Produce a single shell command to run MVR.
}
