% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.cmd.R
\name{cmd.phydstar}
\alias{cmd.phydstar}
\title{Produce a single PhyD* shell command.}
\usage{
cmd.phydstar(infile.d, infile.v = NA, outfile = NA, pr = PR.PHYD,
  method = "BioNJ", fs = 15, binary = TRUE,
  negative.branch.length = FALSE, lower.triangular = TRUE)
}
\value{
Character string
}
\description{
Produce a single PhyD* shell command.
}
