% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.cmd.R
\name{cmd.rm.resistance}
\alias{cmd.rm.resistance}
\title{Produce the shell command that masks resistance mutations.}
\usage{
cmd.rm.resistance(indir, infile, outfile, outdir = indir,
  prog = PR.RM.RESISTANCE, verbose = 1)
}
\arguments{
\item{indir}{Directory of the sequences in file \code{infile}}

\item{infile}{File name to an DNAbin matrix. The file name must end in \code{.R}}

\item{outdir}{Output directory for the bootstrap calculations}

\item{verbose}{Flag to run function in verbose mode.}

\item{alignment.start}{Number that specifies the position of the first nucleotide relative to the HXB2 reference sequence}
}
\value{
Character string
}
\description{
Internal code.
}

