% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cmd.strip.gaps}
\alias{cmd.strip.gaps}
\title{Procude a shell command to strip gaps from alignment.}
\usage{
cmd.strip.gaps(infile, outfile = infile, strip.max.len = NA,
  strip.pc = NA, prog = PR.STRIP.GAPS)
}
\value{
Character string
}
\description{
Internal code.
}

