% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.cmd.R
\name{cmd.treedater}
\alias{cmd.treedater}
\title{Generate commands to date trees with TreeDater.}
\usage{
cmd.treedater(infile.tree, infile.dates, control = list(outfile =
  gsub(".newick$", "_dated.newick", infile.tree), ali.length = 1000),
  pr = PR.TREEDATER)
}
\value{
Character string
}
\description{
Generate commands to date trees with TreeDater.
}
