% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.fun.R
\name{seq.align.based.on.common.reference}
\alias{seq.align.based.on.common.reference}
\title{Merge two alignments with common reference}
\usage{
\method{seq}{align.based.on.common.reference}(in.s, sq,
  return.common.sites = TRUE, regexpr.reference = "HXB2",
  regexpr.nomatch = "-|\\\\?")
}
\arguments{
\item{in.s}{first sequence alignment in **matrix** DNAbin format (ape package)}

\item{sq}{second sequence alignment in **matrix** DNAbin format (ape package)}

\item{return.common.sites}{Flag if the merged alignment should be intersection of the input alignment (TRUE, default) or the union (FALSE).}

\item{regexpr.reference}{Regular expression to identify the reference taxon in the rownames of the input alignments.}

\item{regexpr.nomatch}{Regular expression to identify non-nucleotide characters that should not be matched against each other in the reference. By default, the characters '-' '?' are not mached.}
}
\value{
Merged sequence alignment in matrix DNAbin format (ape package)
}
\description{
Merge two alignments by adding gaps to both alignments in such a way that the common reference in both alignments are aligned with each other.
}
