% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.fun.R
\name{seq.blast.read}
\alias{seq.blast.read}
\title{Read BLAST output}
\usage{
\method{seq}{blast.read}(file, sep = "\\t")
}
\description{
slight modification of read.blast() in pkg RFLPtools; expects blast was run with -outfmt 6
}
