% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seq.find}
\alias{seq.find}
\title{Find the row numbers in a sequence matrix with a particular sequence pattern}
\usage{
\method{seq}{find}(char.matrix, pos0 = NA, from = c(), verbose = 1)
}
\description{
Find the row numbers in a sequence matrix with a particular sequence pattern
}

