% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seq.proportion.ambiguous}
\alias{seq.proportion.ambiguous}
\title{Compute the proportion of ambiguous nucleotides}
\usage{
seq.proportion.ambiguous(seq.DNAbin.mat, exclude = c("-", "?"))
}
\description{
Compute the proportion of ambiguous nucleotides
}

