% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seq.read.GenBank}
\alias{seq.read.GenBank}
\title{Read sequences from GenBank}
\usage{
seq.read.GenBank(access.nb, seq.names = access.nb, species.names = TRUE,
  gene.names = FALSE, as.character = FALSE, attributes = c("origin"))
}
\description{
Read sequences from GenBank
}

