% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.fun.R
\name{seq.rmgaps}
\alias{seq.rmgaps}
\title{Remove gaps in sequence alignment}
\usage{
\method{seq}{rmgaps}(seq.DNAbin.matrix, rm.only.col.gaps = 1,
  verbose = 0)
}
\description{
Remove gaps in sequence alignment
}
