% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seq.write.dna.nexus}
\alias{seq.write.dna.nexus}
\title{Write a NEXUS file}
\usage{
\method{seq}{write.dna.nexus}(seq.DNAbin.mat, ph = NULL, file = NULL,
  nexus.format = "DNA", nexus.gap = "-", nexus.missing = "?",
  nexus.interleave = "NO")
}
\description{
Write a NEXUS file
}

