% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.fun.R
\name{seq.write.dna.phylip.triangular}
\alias{seq.write.dna.phylip.triangular}
\title{Write a triangular phylip file}
\usage{
\method{seq}{write.dna.phylip.triangular}(wd, file = NA)
}
\description{
Write a triangular phylip file
}
