% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.fun.R
\name{seq.find}
\alias{seq.find}
\title{Find the row numbers in a sequence matrix with a particular sequence pattern}
\usage{
\method{seq}{find}(char.matrix, pos0 = NA, from = c(), verbose = 1)
}
\description{
Find the row numbers in a sequence matrix with a particular sequence pattern
}
