% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigphylo.fun.R
\name{seq.write.dna.nexus}
\alias{seq.write.dna.nexus}
\title{Write a NEXUS file}
\usage{
\method{seq}{write.dna.nexus}(seq.DNAbin.mat, ph = NULL, file = NULL,
  nexus.format = "DNA", nexus.gap = "-", nexus.missing = "?",
  nexus.interleave = "NO")
}

